Now downloads playlist to Downloads folder
The program now downloads the playlist to the user's Downloads folder in a new folder with a similar name as the playlist. The program now removes emojis and other characters not allowed in folder names from the playlist title so this is possible.

Moved some helper functions to a separate helper.py file.